terralogic-chatbot-main/requirements.txt
langgraph==0.6.6
langchain==0.3.27
langchain-core==0.3.75
langchain-community==0.3.29
python-dotenv==1.1.1
langchain-google-community[gmail]==2.0.7
ipython==9.4.0
PyMuPDF==1.26.6
sentence-transformers==5.1.0
python-docx==1.2.0
faiss-cpu==1.12.0
langgraph-checkpoint-sqlite==2.0.11
tiktoken==0.11.0
langchain-openai==0.3.32
langchain-text-splitters==0.3.10
pyyaml==6.0.1
beautifulsoup4==4.12.3
Flask-Cors==6.0.1
Flask==3.1.2
gunicorn==23.0.0
email-validator==2.3.0
dnspython==2.8.0
Werkzeug==3.1.3
APScheduler==3.11.0
pandas==2.3.2
bleach==6.2.0
google-cloud-storage==3.4.0
msgpack==1.1.1
boto3==1.35.95
langchain_experimental==0.3.4
Markdown==3.8.2
mdx-truly-sane-lists==1.3
markdown-link-attr-modifier==0.2.1
